set up basic modal for edit tags
